upgrade ffmpge lib

+++ myplayer/src/main/cpp/XsAudio.cpp
@@ -196,7 +196,7 @@ void XsAudio::initOpenSLES() {
     SLDataFormat_PCM pcm = {
             SL_DATAFORMAT_PCM,//播放pcm格式的数据
             2,//2个声道（立体声）
-            getSampleRateForOpenSLES(sample_rate),//44100hz的频率
+            (uint32_t)getSampleRateForOpenSLES(sample_rate),//44100hz的频率
             SL_PCMSAMPLEFORMAT_FIXED_16,//位数 16位
             SL_PCMSAMPLEFORMAT_FIXED_16,//和位数一致就行
             SL_SPEAKER_FRONT_LEFT | SL_SPEAKER_FRONT_RIGHT,//立体声（前左前右）
